requirements.txt
langchain
langchain-openai
langchain-core
langchain_community
docx2txt
langchain-openai
langchain-core
langchain_community
docx2txt
pypdf
langchain_chroma
python-multipart
streamlit
langchain-groq
ipykernel
sentence_transformers
python-dotenv
unstructured
huggingface_hub
langchain-huggingface
langchain-ollama
fastapi
uvicorn
pydantic
streamlit
langchain_chroma
python-multipart
streamlit